[Feature] Generate enemies (#23)
* spawn coordinates

* first tentative of implementation

* working auto-Spawn

* delete sprite on delete

* Update ProjectileFireball.cpp

+++ assets/Resources/main/textures/TileMap.cpp
@@ -17,7 +17,7 @@ TileMap::~TileMap() {
         delete map;
 }
 
-void TileMap::moveMap(int increment) {
+void TileMap::moveMap(float increment) {
     position += increment;
     collision->changePositionRelative(glm::vec2(increment, 0));
     render();
